requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
oauthlib==3.2.0
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tweepy==3.8.0
urllib3==1.26.11
